[!!!] Deprecate AbstractWidget
The abstract widget only contains a helper method
to convert a flexform string to an array, which
really should be done in the widget settings model already.

Deprecate the abstract and add a method to the model
to return the array directly.

This is breaking for widgets, that did not use the
"settings" namespace in their flexforms