[FEATURE] Create basic JS Api for interactive widgets
Let widgets define a AMD module that will be initialized
with each widget that is rendered.

The AMD module can be registered when registering the widget
with the "jsModule" property.

The module must implement bootstrapWidget, which will be
called with the widget DOM element and the dashboard app
on initial rendering of the widget.

We also introduce events when the widget content will
be and has been updated.